// front/src/types/api.d.ts
export interface components {
    schemas: {
        /** CurrentUserResponse */
        CurrentUserResponse: {
            email: string;
            id: string;
            token: string;
            username: string;
        };
        /** LoginRequest */
        LoginRequest: {
            email_or_username: string;
            password: string;
            remember_me?: boolean;
        };
        /** RegisterError */
        RegisterError: {
            errors: {
                email?: string[];
                password?: string[];
                username?: string[];
            };
        };
        /** RegisterRequest */
        RegisterRequest: {
            email: string;
            password: string;
            username: string;
        };
        /** UpdateEmailError */
        UpdateEmailError: {
            errors: {
                current_password?: string[];
                email?: string[];
            };
        };
        /** UpdateEmailRequest */
        UpdateEmailRequest: {
            current_password: string;
            email: string;
        };
        /** UpdatePasswordError */
        UpdatePasswordError: {
            errors: {
                current_password?: string[];
                password?: string[];
            };
        };
        /** UpdatePasswordRequest */
        UpdatePasswordRequest: {
            current_password: string;
            password: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}